Rafactor Permission ID, change type of Duration

+++ app/src/at-daiho/java/asiantech/internship/summer/model/Song.kt
@@ -5,6 +5,6 @@ import android.net.Uri
 data class Song(
         var title: String,
         var artist: String,
-        var duration: Long,
+        var duration: Int,
         var cover: Uri
 )
